+++ src/main/java/com/heftyb/inventorykeeper/repositories/InventoryItemRepository.java
@@ -3,5 +3,5 @@ package com.heftyb.inventorykeeper.repositories;
 import com.heftyb.inventorykeeper.models.InventoryItem;
 import org.springframework.data.repository.CrudRepository;
 
-public interface InventoryItemRepository extends CrudRepository<InventoryItem, Integer> {
+public interface InventoryItemRepository extends CrudRepository<InventoryItem, Long> {
 }
